requirements/dev.txt
pyramid_debugtoolbar==2.3
nose==1.3.6
webtest==2.0.18
coverage==3.7.1
flake8==2.4.0
sphinx==1.3.1
tox==1.9.2
